tests/wpt/web-platform-tests/html/browsers/windows/browsing-context-names/browsing-context-choose-parent.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: Choose browsing context - the given name is '_parent'</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="parent1.html" name="parentWin" style="display:none"></iframe>
<script>

var t = async_test("The parent browsing context must be chosen if the given name is '_parent'");

window.addEventListener("message", function (e) {
  t.step(function() {
    assert_equals(e.data.name, "parentWin", "The browsing context name should be 'parentWin'.");
  });
  t.done();
}, false);

</script>
